backend/requirements.txt
flask==2.2.5
flask-cors==3.0.10
python-dotenv==1.0.1
openai==1.14.0
anyio>=4.4.0,<5.0.0